scx_bpfland: Support balance_power and balance_performance modes
Use the following logic when running in auto mode (`-m auto`):

 +------------------------+----------------+
 | Current Power Profile  | Active Cores   |
 +------------------------+----------------+
 | powersave              | slow cores     |
 | balance_power          | slow cores     |
 | balance_performance    |  all cores     |
 | performance            | fast cores     |
 +------------------------+----------------+

With this logic, on battery-powered systems, the scheduler prioritizes
the slow cores when running on battery. When plugged into AC power, it
uses all cores more evenly or it prioritizes the fast cores if the
"performance" power profile is active.

If the CPU frequency scaling driver supports only a "balanced" mode, the
scheduler operates as balance_performance, using all cores equally.

Signed-off-by: Andrea Righi <arighi@nvidia.com>